requirements.txt
# taken from https://github.com/simplescaling/s1/blob/main/requirements.txt
openai==1.56.1
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
transformers==4.46.1
wandb==0.17.3
datasets==3.1.0
accelerate==1.0.1
vllm==0.6.4.post1
ipykernel==6.28.0
ipython==8.20.0
gradio==4.44.0
trl==0.12.0
lingua-language-detector==2.1.0
stanza==1.10.1
fasttext==0.9.3